Check for duplicated 'text' fields before plotting
Checks for duplicated 'text' fields before plotting. If
duplicates are found, a further determination is performed to
get the duplicated values to report them back to the user.
This 2-step process is done for performance reasons: checking for
duplicates is quick, getting their actual values can be time
consuming if the data is very large.

Since jQuery is already being used, a jQuery dialog is presented
to display duplicate information.

With this check in place, updateSelectedPointColors has been
updated to only check the 'text' field. ('text' is what is displayed
on the tooltip when hovering over a point).